+++ api/controllers/EventController.js
@@ -8,10 +8,12 @@ module.exports = {
       const {type} = req.query
       const order = 'DESC';
       let where = {
+        publish: true,
         type: ["internal-event", "external-event"]
       }
 
-      const posts = await Post.findAllHasJoin({order, where});
+      let posts = await Post.findAllHasJoin({order, where});
+
       const social = SocialService.forPost({posts});
       const items = posts;
       const data = {items}
@@ -27,8 +29,8 @@ module.exports = {
       const {id, name} = req.params
       let data = await Post.findByIdHasJoinByEvent({id, name});
 
-      if(!data){
-        sails.log.error(`Event ID or Name: ${id || name} ,data not found.`);
+      if(!data || !data.publish){
+        sails.log.error(`Event ID or Name: ${id || name} ,data not found or not publish.`);
         return res.notFound();
       }
 
@@ -125,6 +127,7 @@ module.exports = {
         email,
         note,
         token,
+        productionStatus: event.price === 0 ? 'PAID' : 'NEW',
       }).catch(sequelize.UniqueConstraintError, function(err) {
         throw Error('此交易已失效，請重新下訂')
       });
@@ -148,12 +151,46 @@ module.exports = {
       if (event.signupCount > event.limit) {
         throw Error('票卷已賣完');
       }
-      sails.log.warn('歐付寶訂單建立完成 EventOrder');
-      // transaction.commit();
-      return res.view({
-        AioCheckOut: AllpayService.getPostUrl(),
-        ...allPayData.config
-      });
+
+      if (event.price === 0) {
+
+        allPayData.allpay.RtnMsg = '免費活動';
+        allPayData.allpay.ShouldTradeAmt = 0;
+        allPayData.allpay.TradeAmt = 0;
+        allPayData.allpay.PaymentType = '免費活動';
+        allPayData.allpay.PaymentDate = moment(new Date()).format("YYYY/MM/DD");
+        await allPayData.allpay.save();
+        // transaction.commit();
+
+        await event.save();
+
+        try {
+          let messageConfig = {};
+          messageConfig.serialNumber = MerchantTradeNo;
+          messageConfig.email = email;
+          messageConfig.username = user.displayName;
+          sails.log.debug(messageConfig);
+          messageConfig = await MessageService.eventPaymentConfirm(messageConfig);
+          const message = await Message.create(messageConfig);
+          await MessageService.sendMail(message);
+          sails.log.warn('到店購買訂單建立完成 RecipeOrder 寄送 Email id:', message.id);
+        } catch (e) {
+          sails.log.error('寄信失敗', e)
+        }
+
+        return res.redirect(`/event/done?t=${MerchantTradeNo}`);
+
+      } else {
+
+        sails.log.warn('歐付寶訂單建立完成 EventOrder');
+        // transaction.commit();
+        return res.view({
+          AioCheckOut: AllpayService.getPostUrl(),
+          ...allPayData.config
+        });
+
+      }
+
     } catch (e) {
       // transaction.rollback();
       sails.log.error('訂單建立 EventOrder 失敗', e.toString());