+++ api/controllers/api/admin/QuoteController.js
@@ -7,9 +7,16 @@ module.exports = {
       const modelName = req.options.controller.split("/").reverse()[0];
       let result;
       if (serverSidePaging) {
-        result = await PagingService.process({query, modelName});
+        const include = {
+          model: Image
+        }
+        result = await PagingService.process({ query, modelName, include });
       } else {
-        const items = await sails.models[modelName].findAll();
+        const items = await sails.models[modelName].findAll({
+          include:{
+            model: Image
+          }
+        });
         result = { data: { items } };
       }
       res.ok(result);
@@ -21,7 +28,14 @@ module.exports = {
   findOne: async (req, res) => {
     try {
       const { id } = req.params;
-      const item = await Slogan.findById(id);
+      const item = await Quote.findOne({
+        where:{
+          id
+        },
+        include:{
+          model: Image
+        }
+      });
       res.ok({data: {item}});
     } catch (e) {
       res.serverError(e);
@@ -31,7 +45,7 @@ module.exports = {
   create: async (req, res) => {
     try {
       const data = req.body;
-      const item = await Slogan.create(data);
+      const item = await Quote.create(data);
       let message = 'Create success.';
       res.ok({ message, data: { item } } );
     } catch (e) {
@@ -44,7 +58,7 @@ module.exports = {
       const { id } = req.params;
       const data = req.body;
       const message = 'Update success.';
-      const item = await Slogan.update(data ,{
+      const item = await Quote.update(data ,{
         where: { id, },
       });
       res.ok({ message, data: { item } });
@@ -56,7 +70,7 @@ module.exports = {
   destroy: async (req, res) => {
     try {
       const { id } = req.params;
-      const item = await Slogan.deleteById(id);
+      const item = await Quote.deleteById(id);
       let message = 'Delete success';
       res.ok({message, data: {item}});
     } catch (e) {