+++ api/models/Message.js
@@ -21,7 +21,29 @@ module.exports = {
     text: Sequelize.TEXT,
     html: Sequelize.TEXT,
     success: Sequelize.BOOLEAN,
-    error: Sequelize.STRING
+    error: Sequelize.STRING,
+    
+    createdDateTime:{
+      type: Sequelize.VIRTUAL,
+      get: function(){
+        try{
+          return UtilsService.DataTimeFormat(this.getDataValue('createdAt'));
+        } catch(e){
+          sails.log.error(e);
+        }
+      }
+    },
+
+    updatedDateTime:{
+      type: Sequelize.VIRTUAL,
+      get: function(){
+        try{
+          return UtilsService.DataTimeFormat(this.getDataValue('updatedAt'));
+        } catch(e){
+          sails.log.error(e);
+        }
+      }
+    }
   },
   associations: () => {
   },