+++ migration/changelog.md
@@ -82,4 +82,23 @@
 
 ## v6
 
+
 ### 新增 verification Email Token
+
+### Post 欄位增加
+* Post 增加 publish 型態 BOOLEAN，預設 true
+
+### Slogans 表格移除
+
+### 新增 Quotes 表格
+
+### RecipeOrder 欄位資料變動
+* productionStatus 增加 "PENDING"
+
+### EventOrder 欄位資料變動
+* productionStatus 增加 "PENDING"
+
+
+### Feed 欄位變動
+* 修改 fullPicture 欄位 型態 STRING 修改為 TEXT
+* type 欄位等於 'status'，將 publish欄位 修改為 false